+++ spec/dynamic_config_spec.rb
@@ -15,7 +15,9 @@ describe Anyway::Config, type: :config do
     data = Anyway::Config.for(:my_app)
     expect(data[:test]).to eq 1
     expect(data[:name]).to eq "my_app"
-    expect(data[:secret]).to eq "my_secret" if Rails.application.respond_to?(:secrets)
+    unless NOSECRETS
+      expect(data[:secret]).to eq "my_secret"
+    end
     expect(data[:credo]).to eq "my_credo" if Rails.application.respond_to?(:credentials)
   end
 
@@ -47,7 +49,7 @@ describe Anyway::Config, type: :config do
       data = Anyway::Config.for(:my_app)
       expect(data[:test]).to eq 1
       expect(data[:name]).to eq "my_app"
-      expect(data[:secret]).to eq "my_secret" if Rails.application.respond_to?(:secrets)
+      expect(data[:secret]).to eq "my_secret" unless NOSECRETS
       expect(data[:credo]).to eq "my_credo" if Rails.application.respond_to?(:credentials)
       expect(data[:credo_local]).to eq "betheone" if Rails.application.respond_to?(:credentials)
     end